requirements.txt
tensorflow==2.1.0
tqdm
google-cloud-storage
